Split library code from editor command line
